Extract analysis tools; complete tools/ subpackage split
Move multi_attempt, run_code, verify_theorem, minimal_hypotheses, and
profile_proof into tools/analysis.py. All 23 @mcp.tool handlers now live in
the tools/ subpackage; server.py holds the core (mcp, AppContext, lifespan,
BuildCoordinator, shared helpers) and re-exports the handlers. No functional
change.
